// src/interfaces/authInterfaces.tsx
export interface LoginResponse {
    codigo:        string | null;
    mensaje:       string | null;
    usuario:          Usuario;
    mensaje_error: string | null;
    token:         string | null;
}

export interface Usuario {
    id:                   number | null;
    usuario:              string | null;
    nombre:               string | null;
    fecha_creacion:       string | null;
    apellido:             string | null;
    correo:               string | null;
    id_institucion:       number | null;
    estado:               string | null;
    foto:                 string | null;
    password:             string | null;
    last_access:          string | null;
    foto_banner:          string | null;
    id_persona:           number | null;
    password2:            string | null;
    id_emprendedor:       number | null;
    id_usuario:           number | null;
    url_foto:             string | null;
    url_foto_banner:      string | null;
    id_asistente_tecnico: number | null;
    id_mentor:            number | null;
    emprendedor:          number | null;
    mesa_servicio:        number | null;
    mentor:               number | null;
    asistencia_tecnica:   number | null;
    administrador:        number | null;
    supervisorProveedor:  number | null;
    asesor:               number | null;
    inscripciones:        Inscripcion[];
}

export interface Inscripcion {
    id_inscripcion:              number | null;
    estado_inscripcion:          string | null;
    fase:                        number | null;
    id_emprendimiento:           number | null;
    nombre_comercial:            string | null;
    nombre_emprendimiento:       string | null;
    ruc_emprendimiento:          string | null;
    razon_social_emprendimiento: string | null;
}